+++ AWSAppSyncClient/AWSAppSyncClient.swift
@@ -1,6 +1,16 @@
 //
-//  AWSAppSyncClient.swift
-//  AWSAppSyncClient
+// Copyright 2010-2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License").
+// You may not use this file except in compliance with the License.
+// A copy of the License is located at
+//
+// http://aws.amazon.com/apache2.0
+//
+// or in the "license" file accompanying this file. This file is distributed
+// on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+// express or implied. See the License for the specific language governing
+// permissions and limitations under the License.
 //
 
 import Foundation
@@ -17,6 +27,8 @@ public protocol ConnectionStateChangeHandler {
 }
 
 public typealias SubscriptionResultHandler<Operation: GraphQLSubscription> = (_ result: GraphQLResult<Operation.Data>?, _ transaction: ApolloStore.ReadWriteTransaction?, _ error: Error?) -> Void
+
+public typealias DeltaQueryResultHandler<Operation: GraphQLQuery> = (_ result: GraphQLResult<Operation.Data>?, _ transaction: ApolloStore.ReadWriteTransaction?, _ error: Error?) -> Void
 
 public typealias OptimisticResponseBlock = (ApolloStore.ReadWriteTransaction?) -> Void
 
@@ -923,8 +935,83 @@ public class AWSAppSyncClient {
     /// - Parameters:
     ///   - baseQuery: The base query to fetch which contains the primary data.
     ///   - baseQueryResultHandler: Closure that is called when base query results are available or when an error occurs. Every time a sync operation is called, a fetch for the baseQuery from the cache will be done first before initiating any other operations.
+    ///   - deltaQuery: The delta query which fetches data starting from the `lastSync` time. Defaults to a no-op query.
+    ///   - deltaQueryResultHandler: Closure that is called when delta query executes. Defaults to a no-op handler.
+    ///   - callbackQueue: An optional queue on which sync callbacks will be invoked. Defaults to the main queue.
+    ///   - subscriptionWithSyncConfiguration: The sync configuration where the baseQuery sync interval can be specified. (Defaults to 24 hours.)
+    /// - Returns: An object that can be used to cancel the sync operation.
+    public func sync<BaseQuery: GraphQLQuery, DeltaQuery: GraphQLQuery>(baseQuery: BaseQuery,
+                                                                        baseQueryResultHandler: @escaping OperationResultHandler<BaseQuery>,
+                                                                        deltaQuery: DeltaQuery? = nil,
+                                                                        deltaQueryResultHandler: DeltaQueryResultHandler<DeltaQuery>? = nil,
+                                                                        callbackQueue: DispatchQueue = DispatchQueue.main,
+                                                                        subscriptionWithSyncConfiguration: SubscriptionWithSyncConfiguration? = nil) -> Cancellable {
+
+        // The compiler chokes on delegating to `AWSAppSyncClient.sync(baseQuery:baseQueryResultHandler:..)`, so we'll invoke
+        // the final return within this method, at the expense of some code duplication.
+        let resolvedSubscriptionWithSyncConfiguration = subscriptionWithSyncConfiguration ?? SubscriptionWithSyncConfiguration()
+
+        let subscription = EmptySubscription.init()
+        let subscriptionResultHandler: SubscriptionResultHandler<EmptySubscription> = { (_, _, _) in }
+
+        // Safe to force-unwrap here; we only care about the EmptyQuery's `operationString`
+        let resolvedDeltaQuery: DeltaQuery = deltaQuery ?? EmptyQuery() as! DeltaQuery
+        let resolvedDeltaCallback: DeltaQueryResultHandler<DeltaQuery> = deltaQueryResultHandler ?? { (_, _, _) in }
+
+        return AppSyncSubscriptionWithSync<EmptySubscription, BaseQuery, DeltaQuery>(appSyncClient: self,
+                                                                                     baseQuery: baseQuery,
+                                                                                     deltaQuery: resolvedDeltaQuery,
+                                                                                     subscription: subscription,
+                                                                                     baseQueryHandler: baseQueryResultHandler,
+                                                                                     deltaQueryHandler: resolvedDeltaCallback,
+                                                                                     subscriptionResultHandler: subscriptionResultHandler,
+                                                                                     subscriptionMetadataCache: self.subscriptionMetadataCache,
+                                                                                     subscriptionWithSyncConfiguration: resolvedSubscriptionWithSyncConfiguration,
+                                                                                     handlerQueue: callbackQueue)
+    }
+
+    /// Performs a sync operation where a base query is periodically called to fetch primary data from the server based on the syncConfiguration.
+    ///
+    /// - Parameters:
+    ///   - baseQuery: The base query to fetch which contains the primary data.
+    ///   - baseQueryResultHandler: Closure that is called when base query results are available or when an error occurs. Every time a sync operation is called, a fetch for the baseQuery from the cache will be done first before initiating any other operations.
+    ///   - subscription: The subscription which will provide real time updates.
+    ///   - subscriptionResultHandler: Closure that is called when a real time update is available or when an error occurs.
+    ///   - deltaQuery: The delta query which fetches data starting from the `lastSync` time.
+    ///   - deltaQueryResultHandler: Closure that is called when delta query executes.
+    ///   - callbackQueue: An optional queue on which sync callbacks will be invoked. Defaults to the main queue.
+    ///   - subscriptionWithSyncConfiguration: The sync configuration where the baseQuery sync interval can be specified. (Defaults to 24 hours.)
+    /// - Returns: An object that can be used to cancel the sync operation.
+    public func sync<BaseQuery: GraphQLQuery, Subscription: GraphQLSubscription, DeltaQuery: GraphQLQuery>(baseQuery: BaseQuery,
+                                                                                                           baseQueryResultHandler: @escaping OperationResultHandler<BaseQuery>,
+                                                                                                           subscription: Subscription,
+                                                                                                           subscriptionResultHandler: @escaping SubscriptionResultHandler<Subscription>,
+                                                                                                           deltaQuery: DeltaQuery,
+                                                                                                           deltaQueryResultHandler: @escaping DeltaQueryResultHandler<DeltaQuery>,
+                                                                                                           callbackQueue: DispatchQueue = DispatchQueue.main,
+                                                                                                           subscriptionWithSyncConfiguration: SubscriptionWithSyncConfiguration? = nil) -> Cancellable {
+        let subscriptionWithSyncConfiguration = subscriptionWithSyncConfiguration ?? SubscriptionWithSyncConfiguration()
+
+        return AppSyncSubscriptionWithSync<Subscription, BaseQuery, DeltaQuery>(appSyncClient: self,
+                                                                                baseQuery: baseQuery,
+                                                                                deltaQuery: deltaQuery,
+                                                                                subscription: subscription,
+                                                                                baseQueryHandler: baseQueryResultHandler,
+                                                                                deltaQueryHandler: deltaQueryResultHandler,
+                                                                                subscriptionResultHandler: subscriptionResultHandler,
+                                                                                subscriptionMetadataCache: self.subscriptionMetadataCache,
+                                                                                subscriptionWithSyncConfiguration: subscriptionWithSyncConfiguration,
+                                                                                handlerQueue: callbackQueue)
+    }
+
+    /// Performs a sync operation where a base query is periodically called to fetch primary data from the server based on the syncConfiguration.
+    ///
+    /// - Parameters:
+    ///   - baseQuery: The base query to fetch which contains the primary data.
+    ///   - baseQueryResultHandler: Closure that is called when base query results are available or when an error occurs. Every time a sync operation is called, a fetch for the baseQuery from the cache will be done first before initiating any other operations.
     ///   - syncConfiguration: The sync configuration where the baseQuery sync interval can be specified. (Defaults to 24 hours.)
     /// - Returns: An object that can be used to cancel the sync operation.
+    @available(*, deprecated, message: "Use sync(baseQuery:baseQueryResultHandler:deltaQuery:deltaQueryResultHandler:callbackQueue:subscriptionWithSyncConfiguration:)")
     public func sync<BaseQuery: GraphQLQuery>(
         baseQuery: BaseQuery,
         baseQueryResultHandler: @escaping OperationResultHandler<BaseQuery>,
@@ -936,7 +1023,7 @@ public class AWSAppSyncClient {
         let deltaQuery = EmptyQuery.init()
         let deltaCallback: (GraphQLResult<EmptyQuery.Data>?, ApolloStore.ReadTransaction?, Error?) -> Void = { (_, _, _) in
         }
-        
+
         return AppSyncSubscriptionWithSync<EmptySubscription, BaseQuery, EmptyQuery>.init(appsyncClient: self,
                                                                                        baseQuery: baseQuery,
                                                                                        deltaQuery: deltaQuery,
@@ -947,7 +1034,7 @@ public class AWSAppSyncClient {
                                                                                        subscriptionMetadataCache: self.subscriptionMetadataCache,
                                                                                        syncConfiguration: syncConfiguration, handlerQueue: callbackQueue)
     }
-    
+
     /// Performs a sync operation where a delta query is initiated for missed updates and a base query  is used to fetch primary data from the server.
     ///
     /// - Parameters:
@@ -957,6 +1044,7 @@ public class AWSAppSyncClient {
     ///   - deltaQueryResultHandler: Closure that is called when delta query executes.
     ///   - syncConfiguration: The sync configuration where the baseQuery sync interval can be specified. (Defaults to 24 hours.)
     /// - Returns: An object that can be used to cancel the sync operation.
+    @available(*, deprecated, message: "Use sync(baseQuery:baseQueryResultHandler:deltaQuery:deltaQueryResultHandler:callbackQueue:subscriptionWithSyncConfiguration:)")
     public func sync<BaseQuery: GraphQLQuery, DeltaQuery: GraphQLQuery>(
         baseQuery: BaseQuery,
         baseQueryResultHandler: @escaping OperationResultHandler<BaseQuery>,
@@ -964,6 +1052,7 @@ public class AWSAppSyncClient {
         deltaQueryResultHandler: @escaping DeltaQueryResultHandler<DeltaQuery>,
         callbackQueue: DispatchQueue = DispatchQueue.main,
         syncConfiguration: SyncConfiguration = SyncConfiguration.defaultSyncConfiguration()) -> Cancellable {
+
         let subs = EmptySubscription.init()
         let subsCallback: (GraphQLResult<EmptySubscription.Data>?, ApolloStore.ReadTransaction?, Error?) -> Void = { (_, _, _) in
         }
@@ -990,6 +1079,7 @@ public class AWSAppSyncClient {
     ///   - deltaQueryResultHandler: Closure that is called when delta query executes.
     ///   - syncConfiguration: The sync configuration where the baseQuery sync interval can be specified. (Defaults to 24 hours.)
     /// - Returns: An object that can be used to cancel the sync operation.
+    @available(*, deprecated, message: "Use AWSAppSyncClient.sync(baseQuery:baseQueryResultHandler:subscription:subscriptionResultHandler:deltaQuery:deltaQueryResultHandler:callbackQueue:subscriptionWithSyncConfiguration:)")
     public func sync<BaseQuery: GraphQLQuery, Subscription: GraphQLSubscription, DeltaQuery: GraphQLQuery>(
         baseQuery: BaseQuery,
         baseQueryResultHandler: @escaping OperationResultHandler<BaseQuery>,